@@ -7,7 +7,7 @@ public class BruteForcePrimeChecker implements PrimeChecker<Integer> {
     @Override
     public boolean isPrime(Integer number) {
 
-        return number > 2 ? IntStream.range(2, number)
+        return number > 1 ? IntStream.range(2, number)
             .noneMatch(n -> (number % n == 0)) : false;
     }
 
