@@ -6,7 +6,7 @@ public class OptimisedPrimeChecker implements PrimeChecker<Integer> {
 
     @Override
     public boolean isPrime(Integer number) {
-        return number > 2 ? IntStream.rangeClosed(2, (int) Math.sqrt(number))
+        return number > 1 ? IntStream.rangeClosed(2, (int) Math.sqrt(number))
             .noneMatch(n -> (number % n == 0)) : false;
     }
 
