@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import com.baeldung.multipledb.model.product.ProductMultipleDB;
+import com.baeldung.multipledb.model.product.Product;
 
-public interface ProductRepository extends PagingAndSortingRepository<ProductMultipleDB, Integer> {
+public interface ProductRepository extends PagingAndSortingRepository<Product, Integer> {
 
-    List<ProductMultipleDB> findAllByPrice(double price, Pageable pageable);
+    List<Product> findAllByPrice(double price, Pageable pageable);
 }
