@@ -23,7 +23,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.baeldung.multipledb.dao.product.ProductRepository;
-import com.baeldung.multipledb.model.product.ProductMultipleDB;
+import com.baeldung.multipledb.model.product.Product;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes=MultipleDbApplication.class)
@@ -36,22 +36,22 @@ public class ProductRepositoryIntegrationTest {
     @Before
     @Transactional("productTransactionManager")
     public void setUp() {
-        productRepository.save(ProductMultipleDB.from(1001, "Book", 21));
-        productRepository.save(ProductMultipleDB.from(1002, "Coffee", 10));
-        productRepository.save(ProductMultipleDB.from(1003, "Jeans", 30));
-        productRepository.save(ProductMultipleDB.from(1004, "Shirt", 32));
-        productRepository.save(ProductMultipleDB.from(1005, "Bacon", 10));
+        productRepository.save(Product.from(1001, "Book", 21));
+        productRepository.save(Product.from(1002, "Coffee", 10));
+        productRepository.save(Product.from(1003, "Jeans", 30));
+        productRepository.save(Product.from(1004, "Shirt", 32));
+        productRepository.save(Product.from(1005, "Bacon", 10));
     }
 
     @Test
     public void whenRequestingFirstPageOfSizeTwo_ThenReturnFirstPage() {
         Pageable pageRequest = PageRequest.of(0, 2);
 
-        Page<ProductMultipleDB> result = productRepository.findAll(pageRequest);
+        Page<Product> result = productRepository.findAll(pageRequest);
 
         assertThat(result.getContent(), hasSize(2));
         assertTrue(result.stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .allMatch(id -> Arrays.asList(1001, 1002)
                 .contains(id)));
     }
@@ -60,11 +60,11 @@ public class ProductRepositoryIntegrationTest {
     public void whenRequestingSecondPageOfSizeTwo_ThenReturnSecondPage() {
         Pageable pageRequest = PageRequest.of(1, 2);
 
-        Page<ProductMultipleDB> result = productRepository.findAll(pageRequest);
+        Page<Product> result = productRepository.findAll(pageRequest);
 
         assertThat(result.getContent(), hasSize(2));
         assertTrue(result.stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .allMatch(id -> Arrays.asList(1003, 1004)
                 .contains(id)));
     }
@@ -73,11 +73,11 @@ public class ProductRepositoryIntegrationTest {
     public void whenRequestingLastPage_ThenReturnLastPageWithRemData() {
         Pageable pageRequest = PageRequest.of(2, 2);
 
-        Page<ProductMultipleDB> result = productRepository.findAll(pageRequest);
+        Page<Product> result = productRepository.findAll(pageRequest);
 
         assertThat(result.getContent(), hasSize(1));
         assertTrue(result.stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .allMatch(id -> Arrays.asList(1005)
                 .contains(id)));
     }
@@ -86,12 +86,12 @@ public class ProductRepositoryIntegrationTest {
     public void whenSortingByNameAscAndPaging_ThenReturnSortedPagedResult() {
         Pageable pageRequest = PageRequest.of(0, 3, Sort.by("name"));
 
-        Page<ProductMultipleDB> result = productRepository.findAll(pageRequest);
+        Page<Product> result = productRepository.findAll(pageRequest);
 
         assertThat(result.getContent(), hasSize(3));
         assertThat(result.getContent()
             .stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .collect(Collectors.toList()), equalTo(Arrays.asList(1005, 1001, 1002)));
 
     }
@@ -101,12 +101,12 @@ public class ProductRepositoryIntegrationTest {
         Pageable pageRequest = PageRequest.of(0, 3, Sort.by("price")
             .descending());
 
-        Page<ProductMultipleDB> result = productRepository.findAll(pageRequest);
+        Page<Product> result = productRepository.findAll(pageRequest);
 
         assertThat(result.getContent(), hasSize(3));
         assertThat(result.getContent()
             .stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .collect(Collectors.toList()), equalTo(Arrays.asList(1004, 1003, 1001)));
 
     }
@@ -117,12 +117,12 @@ public class ProductRepositoryIntegrationTest {
             .descending()
             .and(Sort.by("name")));
 
-        Page<ProductMultipleDB> result = productRepository.findAll(pageRequest);
+        Page<Product> result = productRepository.findAll(pageRequest);
 
         assertThat(result.getContent(), hasSize(5));
         assertThat(result.getContent()
             .stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .collect(Collectors.toList()), equalTo(Arrays.asList(1004, 1003, 1001, 1005, 1002)));
 
     }
@@ -131,11 +131,11 @@ public class ProductRepositoryIntegrationTest {
     public void whenRequestingFirstPageOfSizeTwoUsingCustomMethod_ThenReturnFirstPage() {
         Pageable pageRequest = PageRequest.of(0, 2);
 
-        List<ProductMultipleDB> result = productRepository.findAllByPrice(10, pageRequest);
+        List<Product> result = productRepository.findAllByPrice(10, pageRequest);
 
         assertThat(result, hasSize(2));
         assertTrue(result.stream()
-            .map(ProductMultipleDB::getId)
+            .map(Product::getId)
             .allMatch(id -> Arrays.asList(1002, 1005)
                 .contains(id)));
     }
