@@ -1,9 +1,14 @@
 package com.baeldung.multipledb.dao.product;
 
+import java.util.List;
+
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-import com.baeldung.multipledb.model.product.ProductMultipleDB;
+import com.baeldung.multipledb.model.product.Product;
 
-public interface ProductRepository extends PagingAndSortingRepository<ProductMultipleDB, Integer> {
+public interface ProductRepository extends PagingAndSortingRepository<Product, Integer> {
 
+
+    List<Product> findAllByPrice(double price, Pageable pageable);
 }
