@@ -1,8 +1,10 @@
 package com.baeldung.r2dbc.repository;
 
-import com.baeldung.r2dbc.model.Player;
-import org.springframework.data.r2dbc.repository.query.Query;
+import org.springframework.data.r2dbc.repository.Query;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
+
+import com.baeldung.r2dbc.model.Player;
+
 import reactor.core.publisher.Flux;
 
 public interface PlayerRepository extends ReactiveCrudRepository<Player, Integer> {
