@@ -1,19 +1,17 @@
 package com.baeldung.web.controller;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 public class SimpleBookRestControllerIntegrationTest {
 
     private MockMvc mockMvc;
-    private static final String CONTENT_TYPE = "application/json;charset=UTF-8";
 
     @BeforeEach
     public void setup() {
@@ -25,7 +23,7 @@ public class SimpleBookRestControllerIntegrationTest {
         this.mockMvc
             .perform(get("/books-rest/42"))
             .andExpect(status().isOk())
-            .andExpect(content().contentType(CONTENT_TYPE))
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON_VALUE))
             .andExpect(jsonPath("$.id").value(42));
     }
 
